Make sure partitions elements are ints

+++ dask/dataframe/shuffle.py
@@ -785,7 +785,7 @@ def collect(p, part, meta, barrier_token):
 
 
 def set_partitions_pre(s, divisions, ascending=True, na_position="last"):
-    partitions = np.zeros_like(s)
+    partitions = np.zeros_like(s, dtype="int32")
     not_null = s.notna().values
     if ascending:
         partitions[not_null] = divisions.searchsorted(s[not_null], side="right") - 1
